Changes for upgrading from MSVC Community 2015 to 2017.  Seems more strict with include files and const methods.

+++ components/esm/loadscpt.cpp
@@ -1,5 +1,7 @@
 #include "loadscpt.hpp"
 
+#include <iostream>
+
 #include "esmreader.hpp"
 #include "esmwriter.hpp"
 #include "defs.hpp"
@@ -83,7 +85,7 @@ namespace ESM
                     if (mData.mScriptDataSize)
                     {
                         mScriptData.resize(mData.mScriptDataSize);
-                        esm.getHExact(&mScriptData[0], mScriptData.size());
+                        esm.getHExact(&mScriptData[0], (int)mScriptData.size()); // suppress compiler warning
                     }
                     else
                         esm.skipHSub();
